--- sonaranalyzer-dotnet/tests/SonarAnalyzer.UnitTest/TestCases/DoNotExposeListT.CSharp9.TopLevel.cs
@@ -1,9 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-
-var x = 1;
-
-public class Bar<T>
-{
-    public List<T> Method1<T>(T arg) => null; // Noncompliant
-}
